+++ frontend/src/pages/SignupPage/hooks/index.ts
@@ -1,1 +1,1 @@
-export { useSignupForm } from '@/pages/SignupPage/hooks/useSignupForm';
+export { useSignupForm } from './useSignupForm';
